+++ src/main/java/ru/nucodelabs/gem/fxmodel/anisotropy/mapper/AnisotropyFxModelMapper.java
@@ -1,11 +1,15 @@
 package ru.nucodelabs.gem.fxmodel.anisotropy.mapper;
 
-import javafx.collections.FXCollections;
-import javafx.collections.ObservableList;
-import org.mapstruct.*;
-import ru.nucodelabs.gem.fxmodel.anisotropy.*;
+import org.mapstruct.Mapper;
+import org.mapstruct.NullValueMappingStrategy;
+import org.mapstruct.ReportingPolicy;
+import ru.nucodelabs.gem.fxmodel.anisotropy.ObservableAzimuthSignals;
+import ru.nucodelabs.gem.fxmodel.anisotropy.ObservableModelLayer;
+import ru.nucodelabs.gem.fxmodel.anisotropy.ObservablePoint;
 import ru.nucodelabs.gem.fxmodel.map.ObservableWgs;
-import ru.nucodelabs.geo.anisotropy.*;
+import ru.nucodelabs.geo.anisotropy.AzimuthSignals;
+import ru.nucodelabs.geo.anisotropy.ModelLayer;
+import ru.nucodelabs.geo.anisotropy.Point;
 import ru.nucodelabs.geo.anisotropy.calc.map.Wgs;
 
 @Mapper(
@@ -13,31 +17,13 @@ import ru.nucodelabs.geo.anisotropy.calc.map.Wgs;
         unmappedTargetPolicy = ReportingPolicy.ERROR,
         nullValueIterableMappingStrategy = NullValueMappingStrategy.RETURN_DEFAULT
 )
-public abstract class AnisotropyFxModelMapper {
-    protected <T> ObservableList<T> createObservableList() {
-        return FXCollections.observableArrayList();
-    }
+public abstract class AnisotropyFxModelMapper extends AnisotropyCommonMapper {
 
     public abstract ObservablePoint toObservable(Point point);
-
-    public abstract void updateObservable(@MappingTarget ObservablePoint observablePoint, Point src);
-
-    public ObservableWgs toObservable(Wgs wgs) {
-        if (wgs == null) {
-            return null;
-        }
-        return new ObservableWgs(wgs.getLongitudeInDegrees(), wgs.getLatitudeInDegrees());
-    }
-
-    @Mapping(target = "isHidden", source = "hidden")
-    public abstract ObservableSignal toObservable(Signal signal);
 
     public abstract ObservableAzimuthSignals toObservable(AzimuthSignals azimuthSignals);
 
     public abstract ObservableModelLayer toObservable(ModelLayer modelLayer);
-
-    @Mapping(target = "isFixed", source = "fixed")
-    public abstract ObservableFixableValue<Double> toObservable(FixableValue<Double> fixableValue);
 
     public abstract Wgs toModel(ObservableWgs observableWgs);
 }
